requirements.txt
torch
numpy
matplotlib
pyyaml
h5py
git+git://github.com/j-friedrich/OASIS.git@master